Remove data and header block parsing/serializing from h3c_frame_parse/h3c_frame_serialize.
Serializing data and header blocks in h3c_frame_serialize required the
entire buffer (data or header block) to be available when calliand
sending the buffer after calling the function. By not serializing large
buffers in h3c_frame_serialize, the user has full control over how the
buffers are serialized.

Parsing data and header blocks in h3c_frame_parse only stored a pointer
to the start of the buffer in the input along with its size. This is
error prone since the input can easily be deallocated leaving the user
with pointers to invalid memory in the returned frame. Instead of
storing a pointer in the frame, we now only store the size of the buffer
in the returned frame. The user can calculate the start of the buffer by
adding the frame's size (without the data or header block) to the input
buffer. Now the user has full control over parsing the data or header
block and can do so incrementally instead of the full data or header
block having to be available at the time of calling h3c_frame_parse.