Modify varint serialize/parse functions to return H3C_ERROR values.
The new style is consistent with the frame serialize/parse functions and
allows determining whether a varint overflows or the buffer is too small
when calling h3c_varint_serialize without having to call
h3c_varint_size.

A new error H3C_ERROR_INCOMPLETE_VARINT was added which is now returned
by h3c_varint_parse when a full varint is not yet available in the
input.