Make logging macros part of the public log.h API again.
The logging macros were previously made private because they implicitly
assumed the existence of a h3c_log_t pointer variable named log which
isn't necessarily the case outside of h3c's source code. By making the
log variable an explicit parameter of the logging macros, they become
generic enough to be part of the public logging API.

By having the logging macros in the public API, we can use them in tests
and executables that depend on h3c as well which saves us from having to
implement logging separately for each of these targets.