Rename H3C_ERROR macro to H3C_THROW.
h3c's error enum is already named H3C_ERROR so it was confusing to have
a macro of the same name. H3C_THROW is a good alternative since
its name indicates that the macro contains a return statement and
execution will not continue beyond that point (even if C doesn't have
exceptions).