Move encoded size calculation out of `encode` functions.
Previously, the encoded size of an object (frame/varint/...) could be
calculated by passing `NULL` as the `dest` argument to any `encode`
function. However, this complicated the implementation of these
functions (many if checks on `dest`) and presented a complicated API to
just calculate the encoded size of an object since all `encode`
functions return an error which has to be checked.

Now, each `encode` function has an `encoded_size` companion function
which returns the encoded size of the object. The new `encoded_sizeÂ`
functions cannot error and take a vastly reduced amount of arguments
which makes them simpler to use compared to the old way of calculating
encoded sizes.