Remove `H3C_THROW` from the public API.
Because of becoming a private macro, the `H3C_` prefix became
unnecessary and as a result `H3C_THROW` has been renamed to `THROW`. The
`log` parameter was also removed in favor of implicitly assuming `log`
to be present when calling `THROW`.

The advantage of making `H3C_THROW` private is that `error.h` and
`log.h` can be completely isolated from each other without one having to
include the other which is required to implement `H3C_THROW`.

`THROW` was also duplicated to a util.h header for qpack-encode and
qpack-decode so those targets can continue using it as before.