Guarantee `encoded_size` of zero when an error occurs.
This commit also refactors `encoded_size` calculation in general.
Instead of incrementing `encoded_size` separately when
encoding/decoding, we keep track of the start of `src`/`dest` in a
`begin` variable at the start of a function and calculate
`encoded_size` when the function ends by subtracting `begin` from the
current value of `src/dest`.

`huffman_symbol_encode` is also simplified as part of this commit.