Change implementation from C to C++.
Development of the client/server necessitated a move to C++ due to the
need for reference counted buffers. While these can be implemented in C,
the usability drops immensely due to the lack of destructors in C. To
keep a consistent API across h3c the already implemented features were
moved to C++ as well.

Because of the move to C++, the fmt library was also added to provide a
C++ replacement for C varargs. The fmt library is only required when
using h3c's logging.