Refactor frame encoding/decoding.
First of all, frame encoding/decoding was adapted to the class based
design. This follows from the effort started with qpack and followed by
varint encoding/decoding.

Second, the `frame` struct itself was changed to make misuse harder. By
making it a class that can't be modified after construction and
inferring the type based on the payload passed to the constructor, it
becomes impossible to construct a frame with a payload that doesn't
match its type.

Finally, the frame type was changed to an enum class and decoding a
frame with an unknown type now results in an error. By implicitly
converting a frame to its type, a frame instance can be used in a switch
statement to switch on its type without having to call `frame::type()`.
An enum class allows a compiler to warn if there are unhandled frame
values which helps in preventing errors.

The usual const and noexcept specifiers were also added.