Optimize `h3c::buffer::slice` for small slices.
Until now, a slice would always result in either an upgrade
from `std::unique_ptr` to `std::shared_ptr` or a reference count
increase on `std::shared_ptr` regardless of the size of the slice.

Now, we check if the slice fits into `h3c::buffer` SSO array and
construct a SSO `h3c::buffer` if this is the case.