Add initial HTTP/3 client/server implementations.
The first thing to note is that these implementations do not concern
themselves with sockets but instead only consume and produce buffers
with binary data. The user is responsible for transferring the HTTP/3
data produced by h3c to a QUIC implementation of his choice and to pass
on data received from the same QUIC implementaton to h3c. This makes h3c
independent of the QUIC implementation used.

Another interesting design choice is the absence of callbacks from the
client/server API's. Instead, two methods (`send` and `recv`) form the
basis of the API's of both the client and server. `send` is called to
produce encoded HTTP/3 data to be passed on to QUIC. `recv` is called
with data received from QUIC and produces HTTP data (headers, body
buffers and errors). A user keeps calling `recv` until all buffer data
has been processed, processing the resulting events one by one.

To start a request (client) or response (server), the `request` (client)
and `response` (server) methods are used. These methods return a handle
object which can be used to add headers and body buffers to the
request/response. The `response` method takes the corresponding request
id as an additional parameter so the server knows on which stream the
response should be sent.

Most of the implementation of the client and server is shared in
`endpoint/stream`. When one of the shared decoders decodes a frame type
that is processed differently by the client and server, it throws the
`unknown` error following which the client/server specific decoders try
to handle the unknown frame themselves.