Introduce `h3c::buffers` and buffer anchors.
Until now, buffering in the client/server implementations was done by
storing the leftover bytes (that couldn't be processed because more
data was needed) of the last buffer and concatenating those with new the
next incoming data buffer. This was inefficient because the entire
incoming data buffer would have to be copied when any amount of data
had to be buffered. However, it was required because all the decoders
only worked on `h3c::buffer` instances which represent contiguous
memory blocks.

To solve this problem, we decided to adapt all decoders so they could
also work on sequences of `h3c::buffer`s aside from single `h3c::buffer`
instances. To accomplish this, we introduce the `h3c::buffers` class
which presents a sequence of `h3c::buffer`'s as a single contiguous
buffer to its user. Every decoder now takes either a `h3c::buffer`
instance or a `h3c::buffers` instance from which to decode data.
Internally, both these methods use the same implementation which is
possible because `h3c::buffer` and `h3c::buffers` provide mostly the
same interface. Sharing the implementation is done by delegating to a
private templated method that expects the interface provided by both
`h3c::buffer` and `h3c::buffers`.

Together with the introduction of `h3c::buffers, we also replaced the
logic used to reset buffers when an error occurs during decoding by
introducing anchors. The problem with the current approach was that we
had to introduce separate error handling macros (DECODE_*) for the
decoders because aside from handling the error they also had to reset
the current buffer to an initial position. To allow all code to use the
same `TRY`/`THROW` macro's, we introduced buffer anchors. An anchor
saves the current position of the passed buffer and resets it to that
position in its destructor unless the `release` method is called. By
only calling `release` when decoding succeeds, a buffer will always
be reset when an error occurs without having to use separate macros
to do so. Both `h3c::buffer` and `h3c::buffers` provide an anchor
implementation which is used by the decoders that require it.