Rework client/server implementations.
The old implementations had several issues:
1. Determining when a `quic::data` instance was fully processed turned
out to be more complicated than expected. It was also up to the user to
determine when a `quic::data` instance was fully processed instead of
the implementation.
2. Each class had to be capable of buffering data to allow reuse of the
same error handling macros.
3. Custom error handling macros were required to be able to buffer data
and transform/ignore error codes. These macros were hard to debug
because it's not possible to step through them with a debugger.
4. The derived classes in the client/server from the shared classes to
handle unknown frame types had to decode the frame themselves, leading
to duplicated code.
5. The `headers` and `body` encoders/decoders required knowledge of the
stream id and decided themselves when the stream was finished which made
their implementations more complicated than needed.
6. Each encoder/decoder had an implicit conversion to its state but when
using this functionality, the resulting code turned out to be hard to
understand because the `state` type of each encoder/decoder is always a
deeply nested type.

The new client/server implementations solve these problems as follows:
1. Instead of determining when a `quic::data` instance is fully
processed, users pass a `quic::data` instance to the `recv` method along
with a lambda that is called with each `event` produced while processing
the `quic::data` instance. Passing a lambda allows the implementation to
determine when the `quic::data` instance is fully processed instead of
the user. This makes the API much easier to use.
2. Buffering is now only done at the stream level
(`stream::control::decoder` and `stream::request::decoder`). Instead of
buffering when an `incomplete` error is encountered, buffers from
incoming `quic::data` instances are added to a `h3c::buffers` instance
which is then passed to all underlying decoders. By pre-emptively
buffering incoming buffers, we don't have to implement buffering in the
error handling macros anymore. Pre-emptively buffering became possible
with the introduction of `h3c::buffers` which makes buffering cheap
because data does not have to be copied into a single contiguous
`h3c::buffer` instance anymore.
3. While buffering did not need to be implemented anymore in the
error handling macros, they still had to transform/ignore errors and
update the class `state_` member to `state::error` in case of a fatal
error. After some detective work, it turned out the error
transform/ignore logic was only required in `stream::request::decoder`
and in a simpler form in `stream::control::decoder` so we moved the
logic out of the error handling macros. The class state change logic was
extracted into a `state_error_handler` class which takes a state type,
a state reference and an error code reference and updates the state to
the error state in its destructor if the error code reference contains a
fatal error. With these changes and manual handling of errors in some
specific places, all custom error handling macros were able to be
replaced with the common `TRY` and `THROW` macros.
4. Instead of delegating to the base class and checking if it returns
`error::unknown`, the base class now declares a pure virtual `process`
method that takes the unknown frame and returns an event. The derived
classes implement this method and the base class calls it whenever an
unknown frame is encountered. This removes duplicate code since decoding
the actual frame is now done in the base class and allowed us to make
the implementation of the base classes completely private, only exposing
the `process` method to the derived classes in order to be implemented.
5. The `body` and `headers` encoders/decoder now return encoded buffers
and `header`/`buffer` instances respectively instead of `quic::data` and
`event` instances. This allows them to work without knowledge of the
stream id and when to set the fin boolean in the `event` instance. This
logic was moved to the request stream instead.
6. Instead of implicit conversions to their state type, all classes now
provide boolean methods that allow checking if they are in a specific
state. Each `state` type is now private and only the state checking
methods that are actually needed are provided to be used by other
classes. The result is more readable code compared to checking the state
of the class directly with the `statee` type.

The request/control stream encoders/decoders were also renamed to
sender/receiver which are more descriptive names since the
request/control streams mostly delegate to actual encoders/decoders
instead of doing any actual encoding/decoding themselves.