Rename to bnl (Binary Networking Library)
In preparation of adding QUIC to the repository, we rename the project
to Binary Networking Library instead of h3c to reflect the repository
not only containing a HTTP/3 implementation but a QUIC implementation as
well.

Instead of a single library, the code has been divided into multiple
libraries: core, log, util and http3. Each library corresponds with a
CMake target and can be depended on by other targets.

The `logger` class was moved to the `log` namespace and renamed to
`api`. The provided logging implementation was renamed to `console`
instead of `fprintf`.

The git tag of fmt was updated to the latest available commit hash to
fix a compile error with VS2019.

All HTTP/3 code now resides in the `http3` namespace. The top level
namespace is now `bnl`.