Move vocabulary data types to core module.
The problem with putting API types such as `http3::event` and
`quic::event` in their respective modules is that to allow using
`quic::event` in the http3 API, the http3 module would have to depend
on the quic module which we want to avoid. Putting the vocabulary
types in the core module avoids this problem since all implementations
depend on the core module already.

This commit also adds `quic::event` as a core vocabulary type for the
upcoming QUIC implementation ands refactors the http3 implementation to
use it.