Remove stream namespace and reintroduce class macros.
The namespace was still defined in the headers and body codecs. Here it
was just a leftover from an earlier refactor and could be removed. In
the endpoint folder, the stream namespace was renamed to the endpoint
namespace to be more consistent with the folder layout.

The macros to reduce deleted/defaulted move and copy constructors were
also reintroduced to reduce on boilerplate. With this change we also
made a lot of types that don't make sense to be copied move only.