Rename `mutable_buffer` and `mutable_buffer_view`.
The new names are `buffer_mut` and `buffer_view_mut` respectively.

While we tend to avoid abbrevatiations, the `mutable` prefix made the
type names too long and was bad for discoverability as well. Now, when
typing `buffer`, the mutable variants will pop up in code completion as
well. We use the `mut` suffix since it's a widely established convention
in Rust and so has somewhat of a chance of being familiar to C++
programmers as well (at least those that have heard of Rust).