Refactor `buffer` and `buffer_view`.
We remove the `static_` optimization for `buffer` since there's no way
to reliably know if a passed pointer points to a static string. Because
static string now have to be copied into the `buffer` instance, we
increase the `SSO_THRESHOLD` to 20 to make sure most of the static
HTTP/3 QPACK indexed header field strings benefit from the SSO.

We also fix clang-tidy warnings for `buffer` and `buffer_view` in this
commit.