Downgrade required CMake version to 3.13.
Debian Buster includes CMake 3.13 in its stable repositories. By
downgrading the minimum required CMake version to 3.13, bnl can be built
on Debian without having to installing a newer CMake version manually.

The only required change was to duplicate the
`FetchContent_MakeAvailable` macro (as `reproc_MakeAvailable`) which was
added in CMake 3.14.