Move `quic::data` and `quic::error` into `quic::event`.
`quic::error` conflicted with the error code of the same name. This
problem was solved by moving both `quic::data` and `quic::error` into
`quic::event`. The `id` and `fin` members were moved out of `quic::data`
and `quic::error` and made members of `quic::event` instead to be
consistent with `http3::event`. This removed the need for separate types
for data and errors and these are now just typedefs to `buffer` and
`std::error_code` respectively.

The rest of the code was refactored to work with `quic::event` instead
of `quic::data`.