Extract lookahead classes from buffer views.
By adding the `consume` and `copy` methods on the `buffer_view` class,
we were putting multiple concepts into a single class:
- A const non-owning view of a memory region of bytes (original
responsibility)
- An iterable buffer class that allows looking ahead in a `bnl::buffer`
without consuming data from it.

The lookahead concept was easily separated into separate classes for
both `bnl::buffer` and `bnl::buffers`. This obsoleted
`bnl::buffers_view` and as a result it was removed.