Refactor decoders to use explicit template instantiation.
Until now, decoders defined overloads of the `decode` method for
`bnl::buffer`, `bnl::buffers`, `bnl::buffer::lookahead` and
`bnl::buffers::lookahead` to limit the types that can be passed to
`decode`.

Now, we use explicit template instantiation to only instantiate the
`decode` method for the types we want it to work on. We also make
lookahead types be constructible from other lookahead types to allow a
single `decode` implementation for both buffer sequence types and
lookahead types.

We use macros to avoid duplicating the decode declarations and
definitions.