Introduce `base::result<T>` for error handling.
Error handling based on `std::error_code` reference passing to methods
was problematic due to two reasons:

1. We had to remember to check the error after each method call. The
compiler did not help us in enforcing this in any way.

2. We had to use a custom `base::nothing` return type for methods
instead of `void` because the `TRY` and `THROW` macros used `return {};`
which doesn't work with `void`.

Now, methods return `result<T>` which contains either a value or an
error. Because the `value` method has to be called to get the value, it
becomes a lot harder to forget to check the error. Further, if the
`value` method is called when the result contains an error, an
`std::system_error` exception is thrown containing the error.

When a function has no return value but does report errors, we now
simply return `std::error_code`. This makes `base::nothing` obsolete and
it is removed as a result.

The `TRY` macro works with methods returning `result<T>` or
`std::error_code` and within methods returning `result<T>` or
`std::error_code`.

Since the new `TRY` macro depends on statement expressions which are
only supported by GCC and Clang, MSVC is now unsupported.

Because the new error handling breaks `state_error_handler`, it and
error states were removed. Using a class after it returns a fatal error
that would put it in an error state is now undefined behaviour.