Simplify `base::buffer` implementation.
Because undo in a buffer isn't supported anymore, there's no point in
storing the position anymore. The one problem is `std::unique_ptr`
doesn't have an aliasing constructor like `std::shared_ptr` does so we
remove the `std::unique_ptr` storage and only use `std::shared_ptr`
storage and small-size-optimization.

Removing the `std::unique_ptr` leaves us with only two types. If we
require that small-size-optimization is used when the size of the buffer
is smaller than the SSO threshold, we can always know what storage our
buffer is using by looking at its size. This allows us to remove the
`type_` member and corresponding enum.

By storing the begin and end pointer of the memory block instead of the
begin pointer and size, consuming from the buffer only needs to
increment the begin pointer instead of incrementing the begin pointer
and decrementing the size.

One caveat is that the `consumed` method had to be removed since it
couldn't be implemented anymore with the new implementation (and as a
result, `base::buffers::consumed` had to be removed as well). However,
it turns out the `consumed` method on these classes wasn't used (only
the `lookahead::consumed` methods are used), so they can be removed
without any issues.