Add `handle` class to allow easy building of requests and responses.
Until now, each wrapper on top of a connection class had to reproduce
the same 4 methods to add headers, end headers, add body parts and end
body. By introducing a handle class, the handle can simply be returned,
reducing the amount of methods to reproduce to a single one. A handle
points to a `stream::request::sender` class. When created with a pointer
of a specific sender, it intrusively sets the handle pointer of the
sender to itself. Both the handle and sender make sure they still point
as each other as both objects move through the code. The handle errors
if the sender has been destroyed.