Big changes:
- Switched to outcome for error handling. Our own result type was
getting out of hand so it made sense to switch to an actually decent
implementation of a result type in C++.

- Switched to proposed system_error2 for representing errors.
`std::error_code` can only store ints which is not sufficient for our
use case. `status_code` gives us a lot more flexibility in the way we
represent errors and works seamlessly with outcome.

- Subdivided protocol errors. QUIC errors consist out of transport level
connection errors and application errors. Application errors consist of
lots of extra state to keep track of and as a result are better handled
in a separate type. Even more so since they are often not an error but
just an indication of what's happening on the other side of the
connection. Application errors are reported as an event via the
`handler` given to the `recv` method of the connection so they can be
passed on to the application layer. Transport errors are returned via
the `recv` method as they are fatal and need special handling.

- Removed CHECK, ASSERT and RAISE macros. ASSERT was hardly used since
it required a logger and CHECK's inverted conditions made code harder
to read compared to the full if statement. RAISE has an equivalent
`assume_value` in outcome.

- Many bugs discovered during interop testing with ngtcp2 fixed.

- Errors that do not need to be logged are now returned directly instead
of via the `THROW` macro.

- Error enums and quic::path moved to the base module since they are
vocabulary types.

- Events changed to not directly have id and fin members. This was
originally the case so we could use `std::error_code` as the error event
which could only store a single int. Now that we don't use
`std::error_code directly in events anymore, we can move the `id` and
`fin` back inside of the payload structs that need them.