Remove util module and per class loggers.
The error macros weren't strictly necessary anymore and the logging
macros could be replaced by simply shortening the public logging macros.

The per class loggers forced everything to be a class even if the logger
was the only member. By making the logger global we can remove
non-trivial amounts of code. Should we wish per connection logging
configuration, we can just make the logger thread_local and have a
connection method to configure the logger to log for that connection.

Removing the per class loggers allowed us to make many of the HTTP/3
codecs functions instead of classes.